now everything is working with custom hooks
